[Feat] comment view 분리
